typo
git-svn-id: svn+ssh://svn.kde.org/home/kde/branches/work/kst@503601 283d02a7-25f6-0310-bc7c-ecb5cbfe19da

+++ viewpaint/kst/kst/kstviewlegend.cpp
@@ -367,7 +367,7 @@ void KstViewLegend::adjustSizeForText(QRect w) {
   } else {
     width = _textWidth + 9*_ascent*_curves.count()/2 - _ascent;
   }
-  double height = _textHeight;
+  int height = _textHeight;
 
   QSize sz(width, height);
